+++ src/deployment/IncomeVaultOwnable2Step.sol
@@ -14,7 +14,7 @@ import {IRuleEngine} from "CMTAT/interfaces/engine/IRuleEngine.sol";
 /* ==== Snapshot === */
 import {ISnapshotSource} from "../interfaces/ISnapshotSource.sol";
 /* ==== IncomeVault === */
-import {IncomeVaultBase} from "../IncomeVaultBase.sol";
+import {IncomeVaultBaseERC2771} from "../IncomeVaultBaseERC2771.sol";
 import {IncomeVaultValidationModule} from "../modules/IncomeVaultValidationModule.sol";
 import {IncomeVaultRestricted} from "../public/IncomeVaultRestricted.sol";
 import {IncomeVaultSnapshotModule} from "../modules/IncomeVaultSnapshotModule.sol";
@@ -38,7 +38,7 @@ import {IIncomeVault} from "../interfaces/IIncomeVault.sol";
 */
 contract IncomeVaultOwnable2Step is
     IncomeVaultValidationModule,
-    IncomeVaultBase,
+    IncomeVaultBaseERC2771,
     Ownable2StepUpgradeable,
     Ownable2StepERC165Module
 {
@@ -46,7 +46,7 @@ contract IncomeVaultOwnable2Step is
     * @param forwarderIrrevocable Address of the forwarder, required for the gasless support
     */
     /// @custom:oz-upgrades-unsafe-allow constructor
-    constructor(address forwarderIrrevocable) IncomeVaultBase(forwarderIrrevocable) {
+    constructor(address forwarderIrrevocable) IncomeVaultBaseERC2771(forwarderIrrevocable) {
         // Disable the possibility to initialize the implementation
         _disableInitializers();
     }
@@ -102,30 +102,30 @@ contract IncomeVaultOwnable2Step is
     //////////////////////////////////////////////////////////////*/
     /* ============ ERC-2771 / Context disambiguation ============ */
     /**
-    * @inheritdoc IncomeVaultBase
+    * @inheritdoc IncomeVaultBaseERC2771
     */
-    function _msgSender() internal view virtual override(IncomeVaultBase, ContextUpgradeable) returns (address sender) {
-        return IncomeVaultBase._msgSender();
+    function _msgSender() internal view virtual override(IncomeVaultBaseERC2771, ContextUpgradeable) returns (address sender) {
+        return IncomeVaultBaseERC2771._msgSender();
     }
 
     /**
-    * @inheritdoc IncomeVaultBase
+    * @inheritdoc IncomeVaultBaseERC2771
     */
-    function _msgData() internal view virtual override(IncomeVaultBase, ContextUpgradeable) returns (bytes calldata) {
-        return IncomeVaultBase._msgData();
+    function _msgData() internal view virtual override(IncomeVaultBaseERC2771, ContextUpgradeable) returns (bytes calldata) {
+        return IncomeVaultBaseERC2771._msgData();
     }
 
     /**
-    * @inheritdoc IncomeVaultBase
+    * @inheritdoc IncomeVaultBaseERC2771
     */
     function _contextSuffixLength()
         internal
         view
         virtual
-        override(IncomeVaultBase, ContextUpgradeable)
+        override(IncomeVaultBaseERC2771, ContextUpgradeable)
         returns (uint256)
     {
-        return IncomeVaultBase._contextSuffixLength();
+        return IncomeVaultBaseERC2771._contextSuffixLength();
     }
 
     /* ============ Access Control ============ */